requirements.txt
# Flask web framework
Flask==1.1.2

# OpenAI API
openai==0.27.0

# Database
pymongo==3.12.0
firebase-admin==5.0.1

# For making HTTP requests
requests==2.26.0

# For handling environment variables
python-dotenv==0.19.1

# For testing
pytest==6.2.5
